git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/branches/JSR-231@245 232f8b59-042b-4e1e-8c03-345bb8c30851

+++ make/stub_includes/win32/jni_md.h
@@ -3,7 +3,7 @@
 #define JNIIMPORT
 #define JNICALL
 
-typedef long jint;
+typedef int jint;
 typedef __int64 jlong;
 typedef signed char jbyte;
